Time: 17 ms (33.41%) | Memory: 56.4 MB (84.00%) - LeetSync

+++ 2581-divide-players-into-teams-of-equal-skill/divide-players-into-teams-of-equal-skill.cpp
@@ -11,7 +11,7 @@ public:
             if(skill[low]+skill[high] != target){
                 return -1;
             }
-            ans+= (long long)skill[low]*skill[high];
+            ans+= skill[low]*skill[high];
             low++;
             high--;
     }
